Re-factor pnl display logic into settings pane
